requirements.txt
azure-cognitiveservices-speech==1.40.0
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyAudio==0.2.14
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
